LC-valid-perfect-square [Runtime: 1842 ms; Memory: 7.6 MB]

+++ binary-search/valid-perfect-square.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPerfectSquare(long num) {
       long n = num/2;
-      for(int i=0 ; i<=num ; i++)
+      for(long i=0 ; i<=num ; i++)
       {
         if(i*i == num) return true;
       }
